fix getProgramsByCustomerId
